+++ spec/rubocop/cli/autocorrect_spec.rb
@@ -90,4 +90,49 @@ RSpec.describe 'RuboCop::CLI --autocorrect' do # rubocop:disable RSpec/DescribeC
       RUBY
     end
   end
+
+  context 'when corrects `RSpec/LetBeforeExamples` with ' \
+          '`RSpec/ScatteredLet`' do
+    before do
+      RuboCop::ConfigLoader
+        .default_configuration
+        .for_all_cops['SuggestExtensions'] = false
+
+      create_file('spec/example.rb', <<~RUBY)
+        RSpec.describe 'Foo' do
+          let(:params) { {} }
+
+          specify do
+            expect(true).to be true
+          end
+
+          let(:attributes) { %i[first_name last_name] }
+        end
+      RUBY
+    end
+
+    it 'rubocop terminates with a success' do
+      expect(cli.run(['-A', '--only',
+                      'RSpec/LetBeforeExamples,' \
+                      'RSpec/ScatteredLet'])).to eq(0)
+    end
+
+    it 'autocorrects be compatible with each other' do
+      cli.run(['-A', '--only',
+               'RSpec/LetBeforeExamples,' \
+               'RSpec/ScatteredLet'])
+
+      expect(File.read('spec/example.rb')).to eq(<<~RUBY)
+        RSpec.describe 'Foo' do
+          let(:params) { {} }
+
+          let(:attributes) { %i[first_name last_name] }
+          specify do
+            expect(true).to be true
+          end
+
+        end
+      RUBY
+    end
+  end
 end